backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
supabase>=2.10.0
PyJWT>=2.8.0
pydantic[email]>=2.0.0
